FossilOrigin-Name: 88ff4e5deab488f855367311b671220e51ffd380f4503fafbe4d68ceb8473fb9

+++ src/winhttp.c
@@ -365,10 +365,10 @@ static void win32_http_request(void *pAppData){
   while( amt<szHdr ){
     if( sslConn ){
 #ifdef FOSSIL_ENABLE_SSL
-      got = ssl_read_server(sslConn, &zBuf[amt], szHdr-amt);
+      got = ssl_read_server(sslConn, &zBuf[amt], szHdr-1-amt, 0);
 #endif
     }else{
-      got = recv(p->s, &zBuf[amt], szHdr-amt, 0);
+      got = recv(p->s, &zBuf[amt], szHdr-1-amt, 0);
       if( got==SOCKET_ERROR ) goto end_request;
     }
     if( got==0 ){
@@ -396,7 +396,7 @@ static void win32_http_request(void *pAppData){
   while( wanted>0 ){
     if( sslConn ){
 #ifdef FOSSIL_ENABLE_SSL
-      got = ssl_read_server(sslConn, zBuf, sizeof(zBuf));
+      got = ssl_read_server(sslConn, zBuf, min(wanted, sizeof(zBuf)), 1);
 #endif
     }else{
       got = recv(p->s, zBuf, sizeof(zBuf), 0);
@@ -439,6 +439,9 @@ static void win32_http_request(void *pAppData){
   in = fossil_fopen(zReplyFName, "w+b");
   fflush(out);
   fflush(aux);
+  if( g.fHttpTrace ){
+    fossil_print("%s\n", zCmd);
+  }
   fossil_system(zCmd);
   if( in ){
     while( (got = fread(zBuf, 1, sizeof(zBuf), in))>0 ){
@@ -467,9 +470,11 @@ end_request:
   /* Make multiple attempts to delete the temporary files.  Sometimes AV
   ** software keeps the files open for a few seconds, preventing the file
   ** from being deleted on the first try. */
-  for(i=1; i<=10 && file_delete(zRequestFName); i++){ Sleep(1000*i); }
-  for(i=1; i<=10 && file_delete(zCmdFName); i++){ Sleep(1000*i); }
-  for(i=1; i<=10 && file_delete(zReplyFName); i++){ Sleep(1000*i); }
+  if( !g.fHttpTrace ){
+    for(i=1; i<=10 && file_delete(zRequestFName); i++){ Sleep(1000*i); }
+    for(i=1; i<=10 && file_delete(zCmdFName); i++){ Sleep(1000*i); }
+    for(i=1; i<=10 && file_delete(zReplyFName); i++){ Sleep(1000*i); }
+  }
   fossil_free(p);
 }
 
@@ -654,8 +659,12 @@ void win32_http_server(
   if( !GetTempPathW(MAX_PATH, zTmpPath) ){
     fossil_panic("unable to get path to the temporary directory.");
   }
-  zTempPrefix = mprintf("%sfossil_server_P%d",
-                        fossil_unicode_to_utf8(zTmpPath), iPort);
+  if( g.fHttpTrace ){
+    zTempPrefix = mprintf("httptrace");
+  }else{
+    zTempPrefix = mprintf("%sfossil_server_P%d",
+                          fossil_unicode_to_utf8(zTmpPath), iPort);
+  }
   fossil_print("Temporary files: %s*\n", zTempPrefix);
   fossil_print("Listening for %s requests on TCP port %d\n",
                (flags&HTTP_SERVER_SCGI)!=0 ? "SCGI" :
